Add GaugeChart test and extract util methods
- Extracted util methods to allow mocking for tests
- Include single lodash methods instead of all to reduce bundle size
- Add initial component test
